Refactor config parsing for each tool
In order to display information to the user about which config file was
used with a given tool, we're separating the config parsing into its own
file. We can decide whether to use the builtin config and also figure
out which config is used if it's not the builtin one.